requirements.txt
# ============================================
# OpenImage - Requirements File
# ============================================

# Core Django
Django>=5.2.6
django-extensions>=3.2.3

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter
dj-database-url>=2.1.0  # Database URL parsing

# Environment & Configuration
python-decouple>=3.8    # Environment variables management

# HTTP & API Clients
requests>=2.31.0        # HTTP library for API calls
urllib3>=2.1.0          # HTTP client

# Image Processing
Pillow>=10.0.0         # Core image library

# AI/ML for Upscaling
torch==2.0.1            # PyTorch (CPU version)
torchvision==0.15.2     # PyTorch vision utilities
# Real-ESRGAN for AI upscaling (optional but recommended)
realesrgan==0.3.0       # AI upscaling
basicsr==1.4.2          # Real-ESRGAN dependency
facexlib==0.3.0         # Face enhancement (Real-ESRGAN dependency)
gfpgan==1.3.8           # Face restoration (Real-ESRGAN dependency)

numpy==1.23.5

scipy>=1.10.0           # Scientific computing (basicsr dependency)
scikit-image>=0.20.0
# Alternative: For GPU support, replace torch lines above with:
# torch>=2.1.0+cu118     # PyTorch with CUDA 11.8
# torchvision>=0.16.0+cu118
# Install with: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

# Image Format Support
opencv-python-headless>=4.8.1   # OpenCV for image processing

# Development Tools
django-debug-toolbar>=4.2.0  # Debug toolbar for development
ipython>=8.18.0              # Enhanced Python shell
ipdb>=0.13.13                # IPython debugger

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0       # Coverage reporting
factory-boy>=3.3.0      # Test fixtures

# Code Quality
black>=23.12.0          # Code formatter
flake8>=6.1.0           # Linter
isort>=5.13.2           # Import sorter
pylint>=3.0.3           # Code analysis
mypy>=1.7.1             # Type checking

# Documentation
mkdocs>=1.5.3                    # Documentation generator
mkdocs-material>=9.5.3           # Material theme
mkdocs-with-pdf>=0.9.3           # PDF export
pymdown-extensions>=10.5         # Markdown extensions
mkdocs-minify-plugin>=0.7.2      # Minification
mkdocs-git-revision-date-localized-plugin>=1.2.2  # Git dates

# API Documentation
drf-yasg>=1.21.7        # Swagger/OpenAPI docs (if using DRF)
django-rest-framework>=0.1.0  # REST framework

# Performance & Caching
redis>=5.0.1            # Redis client (optional)
django-redis>=5.4.0     # Django Redis cache backend
hiredis>=2.3.2          # Fast Redis protocol parser

# Monitoring & Logging
sentry-sdk>=1.39.1      # Error tracking (optional)
python-json-logger>=2.0.7  # JSON logging

# Security
django-cors-headers>=4.3.1  # CORS handling
cryptography>=41.0.7        # Cryptographic recipes

# Utilities
python-dateutil>=2.8.2  # Date utilities
pytz>=2023.3            # Timezone support
humanize>=4.9.0         # Human-readable numbers/dates

# File Handling
python-magic>=0.4.27    # File type detection

# Background Tasks (Optional)
celery>=5.3.4           # Distributed task queue
redis>=5.0.1            # Celery broker/backend

# Production Server
gunicorn>=21.2.0        # WSGI HTTP Server
whitenoise>=6.6.0       # Static file serving

# ============================================
# Optional Dependencies
# ============================================

# For GPU acceleration (uncomment if you have NVIDIA GPU)
# torch>=2.1.0+cu118
# torchvision>=0.16.0+cu118
# Install with: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

# For advanced image analysis
# numpy>=1.24.3
# scipy>=1.11.4

# For ML model optimization
# onnxruntime>=1.16.3

# ============================================
# Notes
# ============================================

# Minimal Installation (without AI upscaling):
# If you don't need Real-ESRGAN upscaling, you can skip:
# - realesrgan
# - basicsr
# - torch
# - torchvision
# - facexlib
# - gfpgan
# - opencv-python
# 
# Install with: pip install -r requirements.txt --no-deps
# Then manually install: Django requests Pillow python-decouple psycopg2-binary

# Docker Installation:
# All dependencies are automatically installed in Docker image

# Windows Users:
# Some packages may require Microsoft Visual C++ Build Tools
# Download from: https://visualstudio.microsoft.com/visual-cpp-build-tools/

# macOS Users:
# Some packages may require Xcode Command Line Tools
# Install with: xcode-select --install

# Linux Users:
# Install system dependencies first:
# sudo apt-get update
# sudo apt-get install python3-dev libpq-dev build-essential